# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Burnaby Activities registration automation API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * active = watching, booked = successfully registered, cancelled = stopped
 */
export type ActivityTargetStatus = typeof ActivityTargetStatus[keyof typeof ActivityTargetStatus];


export const ActivityTargetStatus = {
  active: 'active',
  booked: 'booked',
  cancelled: 'cancelled',
} as const;

export interface ActivityTarget {
  id: number;
  /** e.g. "Swimming" or "Ice Skating" */
  activityName: string;
  /** e.g. "Orca" or "Gliders 2" */
  level: string;
  /**
     * Date registration opens (YYYY-MM-DD)
     * @nullable
     */
  registrationDate?: string | null;
  /**
     * Time to start checking (HH:MM in America/Vancouver)
     * @nullable
     */
  checkWindowStart?: string | null;
  /**
     * Time to stop checking (HH:MM in America/Vancouver)
     * @nullable
     */
  checkWindowEnd?: string | null;
  /** @nullable */
  notes?: string | null;
  /** active = watching, booked = successfully registered, cancelled = stopped */
  status: ActivityTargetStatus;
  /** @nullable */
  lastCheckedAt?: string | null;
  createdAt: string;
}

export interface ActivityTargetInput {
  /** @minLength 1 */
  activityName: string;
  /** @minLength 1 */
  level: string;
  /** YYYY-MM-DD */
  registrationDate?: string;
  /** HH:MM in America/Vancouver, default 09:00 */
  checkWindowStart?: string;
  /** HH:MM in America/Vancouver, default 11:00 */
  checkWindowEnd?: string;
  notes?: string;
}

export type ActivityTargetUpdateStatus = typeof ActivityTargetUpdateStatus[keyof typeof ActivityTargetUpdateStatus];


export const ActivityTargetUpdateStatus = {
  active: 'active',
  booked: 'booked',
  cancelled: 'cancelled',
} as const;

export interface ActivityTargetUpdate {
  /** @minLength 1 */
  activityName?: string;
  /** @minLength 1 */
  level?: string;
  /** @nullable */
  registrationDate?: string | null;
  /** @nullable */
  checkWindowStart?: string | null;
  /** @nullable */
  checkWindowEnd?: string | null;
  /** @nullable */
  notes?: string | null;
  status?: ActivityTargetUpdateStatus;
}

export type BookingLogEntryOutcome = typeof BookingLogEntryOutcome[keyof typeof BookingLogEntryOutcome];


export const BookingLogEntryOutcome = {
  success: 'success',
  failed: 'failed',
  no_spot: 'no_spot',
  scraper_error: 'scraper_error',
  window_closed: 'window_closed',
} as const;

export interface BookingLogEntry {
  id: number;
  /** @nullable */
  targetId?: number | null;
  activityName?: string;
  level?: string;
  outcome: BookingLogEntryOutcome;
  /** @nullable */
  confirmationNumber?: string | null;
  /** @nullable */
  classDate?: string | null;
  /** @nullable */
  classTime?: string | null;
  /** @nullable */
  notes?: string | null;
  attemptedAt: string;
}

export interface RegisteredActivity {
  name: string;
  /** @nullable */
  level?: string | null;
  /** @nullable */
  dates?: string | null;
  /** @nullable */
  times?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  status?: string | null;
}

export type CurrentRegistrationsResultSource = typeof CurrentRegistrationsResultSource[keyof typeof CurrentRegistrationsResultSource];


export const CurrentRegistrationsResultSource = {
  live: 'live',
  cached: 'cached',
  stub: 'stub',
} as const;

export interface CurrentRegistrationsResult {
  registrations: RegisteredActivity[];
  scrapedAt: string;
  source: CurrentRegistrationsResultSource;
  /** @nullable */
  error?: string | null;
}

export type SchedulerTargetStatusSchedulerState = typeof SchedulerTargetStatusSchedulerState[keyof typeof SchedulerTargetStatusSchedulerState];


export const SchedulerTargetStatusSchedulerState = {
  waiting: 'waiting',
  active_window: 'active_window',
  booked: 'booked',
  cancelled: 'cancelled',
} as const;

export interface SchedulerTargetStatus {
  targetId: number;
  activityName: string;
  level: string;
  /** @nullable */
  registrationDate?: string | null;
  /** @nullable */
  checkWindowStart?: string | null;
  /** @nullable */
  checkWindowEnd?: string | null;
  schedulerState: SchedulerTargetStatusSchedulerState;
  /** @nullable */
  nextCheckAt?: string | null;
  /** @nullable */
  lastCheckedAt?: string | null;
}

export interface SchedulerStatus {
  isRunning: boolean;
  targets: SchedulerTargetStatus[];
  checkedAt: string;
}

export interface ScraperStatus {
  configured: boolean;
  hasCredentials: boolean;
  /** @nullable */
  memberId?: string | null;
  /** @nullable */
  siteUrl?: string | null;
}

export interface CheckAndBookInput {
  /** If true, run the full checkout flow but stop before final payment confirmation */
  dryRun?: boolean;
}

export type BookingAttemptResultOutcome = typeof BookingAttemptResultOutcome[keyof typeof BookingAttemptResultOutcome];


export const BookingAttemptResultOutcome = {
  success: 'success',
  failed: 'failed',
  no_spot: 'no_spot',
  scraper_error: 'scraper_error',
  registration_not_open: 'registration_not_open',
  not_configured: 'not_configured',
} as const;

export interface BookingAttemptResult {
  outcome: BookingAttemptResultOutcome;
  message: string;
  /** @nullable */
  confirmationNumber?: string | null;
  /** @nullable */
  classDate?: string | null;
  /** @nullable */
  classTime?: string | null;
  dryRun: boolean;
  /** @nullable */
  logEntryId?: number | null;
}

export interface SchedulerTriggerResult {
  triggered: boolean;
  message: string;
  targetsChecked?: number;
}

export type ListBookingsParams = {
targetId?: number;
limit?: number;
};