# lib/api-zod/src/generated/types/bookingAttemptResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Burnaby Activities registration automation API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingAttemptResultOutcome } from './bookingAttemptResultOutcome';

export interface BookingAttemptResult {
  outcome: BookingAttemptResultOutcome;
  message: string;
  /** @nullable */
  confirmationNumber?: string | null;
  /** @nullable */
  classDate?: string | null;
  /** @nullable */
  classTime?: string | null;
  dryRun: boolean;
  /** @nullable */
  logEntryId?: number | null;
}