# lib/api-zod/src/generated/types/checkAndBookInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Burnaby Activities registration automation API
 * OpenAPI spec version: 0.1.0
 */

export interface CheckAndBookInput {
  /** If true, run the full checkout flow but stop before final payment confirmation */
  dryRun?: boolean;
}