# lib/api-zod/src/generated/types/scraperStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Burnaby Activities registration automation API
 * OpenAPI spec version: 0.1.0
 */

export interface ScraperStatus {
  configured: boolean;
  hasCredentials: boolean;
  /** @nullable */
  memberId?: string | null;
  /** @nullable */
  siteUrl?: string | null;
}